+++ runtime/near-vm-runner/src/imports.rs
@@ -223,7 +223,7 @@ wrapped_imports! {
     keccak512<[value_len: u64, value_ptr: u64, register_id: u64] -> []>,
     ripemd160<[value_len: u64, value_ptr: u64, register_id: u64] -> []>,
     blake2b<[value_len: u64, value_ptr: u64, register_id: u64] -> []>,
-    ecrecover<[hash_ptr: u64, v: u8, r_ptr: u64, s_ptr: u64, register_id: u64] -> []>,
+    ecrecover<[hash_ptr: u64, v: u32, r_ptr: u64, s_ptr: u64, register_id: u64] -> []>,
     // #####################
     // # Miscellaneous API #
     // #####################
